+++ src/Trarizon.Library/Text/TraString.Netstd.cs
@@ -1,15 +1,13 @@
-﻿#if NETSTANDARD
-
-using Trarizon.Library.Numerics;
-
-namespace Trarizon.Library.Text;
+﻿namespace Trarizon.Library.Text;
 public static partial class TraString
 {
+#if NETSTANDARD
+
     public static ReadOnlySpan<char> AsSpan(this string str, Range range)
     {
-        var (ofs, len) = range.GetCheckedOffsetAndLength(str.Length);
+        var (ofs, len) = range.GetOffsetAndLength(str.Length);
         return str.AsSpan(ofs, len);
     }
-}
 
 #endif
+}